rrequirements.txt
streamlit
google-genai>=0.3.0
python-dotenv